requirements.txt
mistune>=3.0,<4.0
python-docx>=1.0,<2.0
requests>=2.28,<3.0
